+++ src/com/landawn/abacus/util/Seq.java
@@ -217,7 +217,7 @@ public final class Seq<T> extends ImmutableCollection<T> {
         }
 
         final Multiset<?> bOccurrences = Multiset.from(b);
-        final List<T> result = new ArrayList<>(N.min(9, size(), b.size()));
+        final List<T> result = new ArrayList<>(N.min(9, size(), N.len(b)));
 
         for (T e : coll) {
             if (bOccurrences.getAndRemove(e) > 0) {
@@ -250,7 +250,7 @@ public final class Seq<T> extends ImmutableCollection<T> {
         }
 
         final Multiset<?> bOccurrences = Multiset.from(b);
-        final List<T> result = new ArrayList<>(N.min(size(), N.max(9, size() - b.size())));
+        final List<T> result = new ArrayList<>(N.min(size(), N.max(9, size() - N.len(b))));
 
         for (T e : coll) {
             if (bOccurrences.getAndRemove(e) < 1) {
@@ -285,7 +285,7 @@ public final class Seq<T> extends ImmutableCollection<T> {
         }
 
         final Multiset<?> bOccurrences = Multiset.from(b);
-        final List<T> result = new ArrayList<>(N.max(9, Math.abs(size() - b.size())));
+        final List<T> result = new ArrayList<>(N.max(9, Math.abs(size() - N.len(b))));
 
         for (T e : coll) {
             if (bOccurrences.getAndRemove(e) < 1) {
@@ -2422,7 +2422,7 @@ public final class Seq<T> extends ImmutableCollection<T> {
      */
     public <U, E extends Exception, E2 extends Exception> List<Pair<T, U>> innerJoin(final Collection<U> b, final Try.Function<? super T, ?, E> leftKeyMapper,
             final Try.Function<? super U, ?, E2> rightKeyMapper) throws E, E2 {
-        final List<Pair<T, U>> result = new ArrayList<>(N.min(9, size(), b.size()));
+        final List<Pair<T, U>> result = new ArrayList<>(N.min(9, size(), N.len(b)));
 
         if (N.isNullOrEmpty(coll) || N.isNullOrEmpty(b)) {
             return result;
@@ -2453,7 +2453,7 @@ public final class Seq<T> extends ImmutableCollection<T> {
      * @see <a href="http://stackoverflow.com/questions/5706437/whats-the-difference-between-inner-join-left-join-right-join-and-ful
      */
     public <U, E extends Exception> List<Pair<T, U>> innerJoin(final Collection<U> b, final Try.BiPredicate<? super T, ? super U, E> predicate) throws E {
-        final List<Pair<T, U>> result = new ArrayList<>(N.min(9, size(), b.size()));
+        final List<Pair<T, U>> result = new ArrayList<>(N.min(9, size(), N.len(b)));
 
         if (N.isNullOrEmpty(coll) || N.isNullOrEmpty(b)) {
             return result;
@@ -2482,7 +2482,7 @@ public final class Seq<T> extends ImmutableCollection<T> {
      */
     public <U, E extends Exception, E2 extends Exception> List<Pair<T, U>> fullJoin(final Collection<U> b, final Try.Function<? super T, ?, E> leftKeyMapper,
             final Try.Function<? super U, ?, E2> rightKeyMapper) throws E, E2 {
-        final List<Pair<T, U>> result = new ArrayList<>(N.max(9, size(), b.size()));
+        final List<Pair<T, U>> result = new ArrayList<>(N.max(9, size(), N.len(b)));
 
         if (N.isNullOrEmpty(coll)) {
             for (T left : coll) {
@@ -2528,7 +2528,7 @@ public final class Seq<T> extends ImmutableCollection<T> {
      * @see <a href="http://stackoverflow.com/questions/5706437/whats-the-difference-between-inner-join-left-join-right-join-and-ful
      */
     public <U, E extends Exception> List<Pair<T, U>> fullJoin(final Collection<U> b, final Try.BiPredicate<? super T, ? super U, E> predicate) throws E {
-        final List<Pair<T, U>> result = new ArrayList<>(N.max(9, size(), b.size()));
+        final List<Pair<T, U>> result = new ArrayList<>(N.max(9, size(), N.len(b)));
 
         if (N.isNullOrEmpty(coll)) {
             for (T left : coll) {
@@ -2655,7 +2655,7 @@ public final class Seq<T> extends ImmutableCollection<T> {
      */
     public <U, E extends Exception, E2 extends Exception> List<Pair<T, U>> rightJoin(final Collection<U> b, final Try.Function<? super T, ?, E> leftKeyMapper,
             final Try.Function<? super U, ?, E2> rightKeyMapper) throws E, E2 {
-        final List<Pair<T, U>> result = new ArrayList<>(b.size());
+        final List<Pair<T, U>> result = new ArrayList<>(N.len(b));
 
         if (N.isNullOrEmpty(b)) {
             return result;
@@ -2691,7 +2691,7 @@ public final class Seq<T> extends ImmutableCollection<T> {
      * @see <a href="http://stackoverflow.com/questions/5706437/whats-the-difference-between-inner-join-left-join-right-join-and-ful
      */
     public <U, E extends Exception> List<Pair<T, U>> rightJoin(final Collection<U> b, final Try.BiPredicate<? super T, ? super U, E> predicate) throws E {
-        final List<Pair<T, U>> result = new ArrayList<>(b.size());
+        final List<Pair<T, U>> result = new ArrayList<>(N.len(b));
 
         if (N.isNullOrEmpty(b)) {
             return result;